Absolutize ignored files so the comparison against the watcher yielded files works
