+++ src/Supervisor.ts
@@ -2,8 +2,8 @@ import type { ChildProcess, StdioOptions } from "child_process";
 import { spawn } from "child_process";
 import { EventEmitter, once } from "events";
 import { setTimeout } from "timers/promises";
-import type { RunOptions } from "./Options.js";
 import type { Project } from "./Project.js";
+import type { RunOptions } from "./ProjectConfig.js";
 import { log } from "./utils.js";
 
 /** */
